+++ src/fpga/matrix_functions.h
@@ -32,7 +32,7 @@ float Sum(
 
 unsigned int SumSizet(
     const cl_mem &mem,
-    size_t size,
+    uint size,
     const cl_context &context,
     const cl_device_id &device);
